requirements.txt
Django>=4.2,<5
django-environ>=0.10
pillow>=9.5
django-money>=3.0.0
price-parser>=0.3.4
jsonschema>=4.17
money>=1.3.0
django-taggit>=4.0.0
django-bootstrap5>=23.3
pymupdf>=1.23.5
Whoosh>=2.7.4
Django-haystack>=3.2.1
djangorestframework>=3.14.0
django-filter>=23.3
django-bootstrap-icons>=0.8.6
zipp>=3.17.0
django-select2>=8.1.2
django-debug-toolbar>=4.2.0
django-mptt>=0.15.0
django-crispy-forms>=2.1
crispy-bootstrap5>=2023.10
django-import-export>=3.3.7
